% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpu_consumerista.R
\name{download_tpu_consumerista}
\alias{download_tpu_consumerista}
\title{Downloads consumerist issues's TPU}
\usage{

  download_tpu_consumerista(regex_law = " ([Ll]ei )?8\\\\.?078((/90)| |$)|CDC|[Cc]ódigo (de )?[Dd]efesa (do )?[Cc]onsumidor| ([Ll]ei )?7\\\\.?347((/85)| |$)")
}
\arguments{
\item{regex_consumerista}{string or regex containing the content to match with the legals subjects in the TPU.}
}
\description{
Downloads consumerist issues's TPU
}
